worker-python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
loguru>=0.7.0
numpy>=1.26.0
sentence-transformers>=3.0.0
transformers==5.2.0
openai>=1.68.0
psycopg[binary]>=3.2.9
psycopg-pool>=3.2.3
